+++ sdk/communication/communication-sms/samples/v1/javascript/sendSmsWithOptions.js
@@ -36,6 +36,7 @@ async function main() {
   const sendOptions = {
     //delivery reports are delivered via EventGrid
     enableDeliveryReport: true,
+    validityPeriodSeconds: 1333,
     //tags are applied to the delivery report
     tag: "marketing"
   };