+++ sdk/communication/communication-sms/CHANGELOG.md
@@ -30,7 +30,7 @@
 ### Breaking Changes
 
 - `SendRequest` renamed to `SmsSendRequest`.
-- `SendOptions` renamed to `SmsSendOptions` and now has an additional field `tag` to add a custom tag to delivery reports (when enabled).
+- `SendOptions` renamed to `SmsSendOptions` and now has an additional field `tag` to add a custom tag to delivery reports (when enabled). Added a field `validityPeriodSeconds` to define maximum delivery report waiting time, its default is 18000 (when delivery report is enabled).
 - `send` no longer returns `RestResponse`, now returns an array of `SmsSendResults`. This contains fields to validate success/failure of each sent message.
 
 ## 1.0.0-beta.3 (2020-11-16)